dialog-whisperer-local/requirements.txt
pytesseract
pyttsx3
mss
Pillow
numpy
sounddevice
keyboard
pytest
pytest-mock
